move tests (ready(), close(), etc) to one file
